requirements.txt
python-dotenv==1.2.2
huggingface-hub==1.10.2
streamlit==1.56.0
